+++ src/autosar/xml/reference.py
@@ -832,3 +832,19 @@ class DiagnosticValueNeedsRef(BaseRef):
     def accepted_sub_types(cls) -> set[ar_enum.IdentifiableSubTypes]:
         """Acceptable values for dest"""
         return {ar_enum.IdentifiableSubTypes.DIAGNOSTIC_VALUE_NEEDS}
+
+
+class FunctionInhibitionNeedsRef(BaseRef):
+    """
+    AR:FUNCTION-INHIBITION-NEEDS--SUBTYPES-ENUM
+    """
+
+    def __init__(self, value: str,
+                 dest: ar_enum.IdentifiableSubTypes = ar_enum.IdentifiableSubTypes.FUNCTION_INHIBITION_NEEDS
+                 ) -> None:
+        super().__init__(value, dest)
+
+    @classmethod
+    def accepted_sub_types(cls) -> set[ar_enum.IdentifiableSubTypes]:
+        """Acceptable values for dest"""
+        return {ar_enum.IdentifiableSubTypes.FUNCTION_INHIBITION_NEEDS}